# More efficient replacement of `NULL` with `NA_integer` in `join_rows()`