--- a/R/join-rows.R
+++ b/R/join-rows.R
@@ -7,7 +7,7 @@ join_rows <- function(x_key, y_key, type = c("inner", "left", "right", "full")) 
   y_loc <- y_split$loc[matches]
 
   if (type == "left" || type == "full") {
-    y_loc <- map(y_loc, function(x) if (is.null(x)) NA_integer_ else x)
+    y_loc <- vec_assign(y_loc, vec_equal_na(matches), list(NA_integer_))
   }
 
   x_loc <- seq_len(vec_size(x_key))